Harden deploy, CI, and public map delivery
