+++ frontend/package.json
@@ -7,6 +7,7 @@
   },
   "scripts": {
     "dev": "vite",
+    "typecheck": "tsc -p tsconfig.json --noEmit",
     "build": "vite build",
     "preview": "vite preview --host"
   },
