+++ nodes/zero2w/ansible/requirements.yml
@@ -1,0 +1,3 @@
+---
+collections:
+  - name: ansible.posix
